The four gates were red and every push this session went over them
`./ops/gates.sh` is the desk's pre-push law, and it had been RED: 27 ruff findings and 39 mypy
errors across 17 files, several of them landed by this session's own earlier commits. A gate
that is red is a gate nobody reads, and this is exactly the "monitoring checks activity, not
completion" failure applied to the gate itself -- the command was run, it printed, and the
verdict at the bottom stopped being the thing that decided whether to push.

Green now, all four: ruff, compileall, pytest --co, mypy 0 errors over 776 source files.

WHAT WAS ACTUALLY WRONG, rather than reformatted:

  libs/research/desk_memory.py   `Any` used in an annotation and never imported. `from __future__
                                 import annotations` is what kept it from raising, so the lesson
                                 injector every organ reads was one `get_type_hints` call from
                                 breaking, invisibly.
  scripts/check_miner_conversion.py  `tested` computed from every hypothesis row and never used
                                 -- a set built and dropped in the module whose whole job is
                                 measuring conversion.
  libs/data/lake.py              the unlayered bucket typed as `dict[str, object]` by inference,
                                 so `+= 1` on its counters was an operator error waiting on a
                                 stricter mypy.
  libs/ops/wiring_audit.py       a nested `if` whose two-statement body I flattened wrongly on
                                 the first attempt, dropping a line; compileall caught it in the
                                 same pass, which is the reason that gate exists separately.

ONE OVERRIDE REPLACES FOUR INLINE IGNORES. `mt5desk` lives under desks/mt5/ and every importer
reaches it through a runtime `sys.path.insert`, so it resolves when the code runs and cannot be
found when mypy looks. Carried inline, ruff's isort wrapped the import and moved the comment off
it, breaking the ignore -- ruff and mypy each demanded the shape the other rejected. A single
`mt5desk.*` override in pyproject ends that loop and silences nothing else.

The rest are annotations and narrowings at sites where a `dict.get` returns `Any | None` and the
value is used as a dict: real ambiguity, made explicit, no behaviour changed.

+++ libs/regime/broker_clock.py
@@ -172,7 +172,7 @@ def estimate_offset(profiles: Mapping[str, Sequence[float]]) -> dict[str, Any]:
     # CIRCULAR MEDIAN, because 23 and 1 are two hours apart and their arithmetic mean is 12.
     ang = np.array(votes, dtype="float64") * (2.0 * math.pi / 24.0)
     mean_ang = math.atan2(float(np.sin(ang).mean()), float(np.cos(ang).mean()))
-    off = int(round((mean_ang * 24.0 / (2.0 * math.pi)) % 24.0)) % 24
+    off = round((mean_ang * 24.0 / (2.0 * math.pi)) % 24.0) % 24
     if off > 12:
         off -= 24                                     # report as a signed offset, not 0..23
     spread = sorted({v if v <= 12 else v - 24 for v in votes})
